+++ packages/native/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @a-little-world/little-world-design-system-native
 
+## 0.11.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @a-little-world/little-world-design-system-core@1.17.0
+
 ## 0.11.2
 
 ### Patch Changes